+++ lib/core/navigation/navigation.dart
@@ -1,0 +1,5 @@
+import 'package:flutter/material.dart' show GlobalKey, NavigatorState;
+
+final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
+
+NavigatorState get nav => navigatorKey.currentState!;