+++ lib/features/auth/presentation/signin_email.dart
@@ -5,7 +5,7 @@ import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:flutter_smart_dialog/flutter_smart_dialog.dart';
 import 'package:lucide_icons_flutter/lucide_icons.dart';
 import 'package:pudding/common/components/btns.dart'
-    show closeBtn, bigTextOnlyBtn;
+    show closeIconBtn, bigTextOnlyBtn;
 import 'package:pudding/common/components/loading_overlay.dart';
 import 'package:pudding/common/components/view_wrappers.dart'
     show layoutBuilder;
@@ -78,7 +78,7 @@ class _SigninEmailPanelState extends ConsumerState<SigninEmailPanel> {
               ),
               Align(
                 alignment: AlignmentGeometry.topRight,
-                child: closeBtn(
+                child: closeIconBtn(
                   onPressed: () => SmartDialog.dismiss(force: true),
                 ),
               ),
@@ -105,13 +105,7 @@ class _SigninEmailPanelState extends ConsumerState<SigninEmailPanel> {
     return Container(
       height: MediaQuery.sizeOf(context).height * heightFactor,
       width: MediaQuery.sizeOf(context).width * widthFactor,
-      decoration: BoxDecoration(
-        color: Colors.white,
-        borderRadius: Parts.defaultBorderRadius,
-        boxShadow: [
-          BoxShadow(color: Colors.grey, blurRadius: 8, spreadRadius: 0.2),
-        ],
-      ),
+      decoration: Parts.defaultBoxDecoration,
       child: Form(
         key: formKey,
         child: Column(
@@ -130,11 +124,11 @@ class _SigninEmailPanelState extends ConsumerState<SigninEmailPanel> {
         child: TextFormField(
           // email field
           controller: emailCtl,
-          decoration: InputDecoration(
+          decoration: const InputDecoration(
             errorMaxLines: 2,
             labelText: 'email',
             // hintText: "pudding@lukecreated.com",
-            icon: const Icon(LucideIcons.mail),
+            icon: Icon(LucideIcons.mail),
           ),
           // TODO: add localized text for email validator
           validator: Validatorless.multiple([