+++ lib/features/collections/data/collection_model.dart
@@ -1,23 +1,34 @@
 import 'package:cloud_firestore/cloud_firestore.dart'
     show DocumentSnapshot, Timestamp;
+import 'package:pudding/features/collections/data/visibility_enum.dart';
 
-// Data model for [TheCollection] document.
+/// Data model for [TheCollection] document.
 class TheCollection {
   final String id;
   final String title;
+  final String description;
   final String ownerId;
-  final List<String> collaborators;
-  // final List<Map<String, Timestamp>> editLogs;
+  final Timestamp createdAt;
+  final Timestamp lastUpdatedAt;
+  final VisibilityEnum visibility;
   final List<String> tags;
 
   TheCollection({
     required this.id,
     required this.title,
+    this.description = "",
     required this.ownerId,
-    required this.collaborators,
-    // required this.editLogs,
+    required this.createdAt,
+    required this.lastUpdatedAt,
+    required this.visibility,
     required this.tags,
   });
+
+  @override
+  String toString() {
+    super.toString();
+    return "TheCollection: ${toFirestore().toString()}";
+  }
 
   factory TheCollection.fromFirestore(
     DocumentSnapshot<Map<String, dynamic>> snapshot,
@@ -26,9 +37,11 @@ class TheCollection {
     return TheCollection(
       id: snapshot.id, // Get the document ID
       title: data['title'],
+      description: data['description'],
       ownerId: data['ownerId'],
-      collaborators: List<String>.from(data['collaborators']),
-      // editLogs: List<Map<String, Timestamp>>.from( data['editLogs'] ).where((it)=> it.),
+      createdAt: data['createdAt'] as Timestamp,
+      lastUpdatedAt: data['lastUpdatedAt'] as Timestamp,
+      visibility: VisibilityEnum.getEnum(data['visibility'].toString()),
       tags: List<String>.from(data['tags']),
     );
   }
@@ -36,9 +49,11 @@ class TheCollection {
   Map<String, dynamic> toFirestore() {
     return {
       'title': title,
+      'description': description,
       'ownerId': ownerId,
-      'collaborators': collaborators,
-      //'editLogs': editLogs,
+      'createdAt': createdAt,
+      'lastUpdatedAt': lastUpdatedAt,
+      'visibility': visibility.name,
       'tags': tags,
     };
   }
@@ -46,17 +61,21 @@ class TheCollection {
   /// returns the [TheCollection] with given values with immutable id
   TheCollection copyWith({
     String? title,
+    String? description,
     String? ownerId,
-    List<String>? collaborators,
-    List<Map<String, Timestamp>>? editLogs,
+    Timestamp? createdAt,
+    Timestamp? lastUpdatedAt,
+    VisibilityEnum? visibility,
     List<String>? tags,
   }) {
     return TheCollection(
       id: id,
       title: title ?? this.title,
+      description: description ?? this.description,
       ownerId: ownerId ?? this.ownerId,
-      collaborators: collaborators ?? this.collaborators,
-      //editLogs: editLogs ?? this.editLogs,
+      createdAt: createdAt ?? this.createdAt,
+      lastUpdatedAt: lastUpdatedAt ?? this.lastUpdatedAt,
+      visibility: visibility ?? this.visibility,
       tags: tags ?? this.tags,
     );
   }